pyproject.toml
[project]
name = "dota-commenting"
version = "0.1.0"
description = "Diploma on HSE Masters"
authors = [
    {name = "Kirill Filatov",email = "filatovka11@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch (>=2.5.0,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
fire = "^0.7.0"


[tool.poetry.group.data-preparation.dependencies]
twitch-dl = "^2.10.0"
moviepy = "^2.1.2"
speechrecognition = "^3.14.1"
faster-whisper = "^1.1.1"
ffmpeg-python = "^0.2.0"

